requirements.txt
python-can>=4.0.0
pyserial>=3.5
